Use a double for the histogram and the total millisecond attribute on the stopwatch
The code was running so quickly on a fast machine the timings of zero 99th, 95th and 50th percentile were useless.

+++ NetClajServer/Metrics/ServerMetrics.cs
@@ -7,7 +7,7 @@ public class ServerMetrics
     public ObservableCounter<long> PacketsProcessed { get; private init;  }
     public ObservableGauge<long> PacketsPerSecond { get; private init; }
     
-    public Histogram<long> PacketProcessHistogram { get; private init; }
+    public Histogram<double> PacketProcessHistogram { get; private init; }
 
     private long _totalPacketsProcessed = 0;
     private long _lastTotalPacketsProcessed = 0;
@@ -29,7 +29,7 @@ public class ServerMetrics
             "pps"
         );
 
-        PacketProcessHistogram = meter.CreateHistogram<long>("Time to process", "ms", "Time to process a packet");
+        PacketProcessHistogram = meter.CreateHistogram<double>("Time to process", "ms", "Time to process a packet");
     }
 
     public void IncrementIncomingPacketsProcessed()
